backend/requirements.txt
Django>=5.2.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-storages>=1.14.2
boto3>=1.34.0
djangorestframework>=3.14.0
drf-yasg>=1.21.7
djangorestframework-simplejwt>=5.3.0
